Allow specifying which profiles cookbook to use in a node attribute.
This allows you to "swap out" the default `cpe_profiles` cookbook for any cookbook by overriding the `profiles_cookbook` attribute. Obviously the named cookbook should be API-compatible.

The only downside I can see here is that we're writing into the node object for a cookbook that we don't explicitly depend on (as in `depends` in the `metadata.rb`. While this technically works this may be bad form from a cookbook/Chef perspective. You woulnd't want to name the depedent cookbook as that would couple this cookbook to that one — which is precisely what is trying to be avoided to have a sort of modular or pluggable (API compatible) cookbook.